Base addtions for index page and topbar fixes
Topbar was too cluttered and the call was made to have only icons for a minimalist look. The initial additon of the index lay has been added aswell. at this point it is just copied from the original design theme. Further along the functionality will be added to dynamically load featured listings and search based on criterias lisetd to trim back the html file.